decoupling and namespace convention
decoupled the ebpf entry point program from the submodule
and changed the class implementation to respect the namespace convention
