+++ .gitignore
@@ -6,6 +6,7 @@
 !.vscode/extensions.json
 *.code-workspace
 
+
 ## Logs ##
 logs/
