+++ include/efimon/meson.build
@@ -60,7 +60,7 @@ install_headers(lib_asm_classifier_headers, subdir : 'efimon/asm-classifier')
 install_headers(lib_logger_headers, subdir : 'efimon/logger')
 install_headers(lib_perf_headers, subdir : 'efimon/perf')
 install_headers(lib_ptrace_capstone_headers, subdir : 'efimon/ptrace-capstone')
-install_headers(lib_sampling_by_pid_headers, subdir : 'efimon/ebpf-modules/sampling-by-pid')
+install_headers(lib_sampling_by_pid_headers, subdir : 'efimon/ebpf-modules/cpu-assembly-sampler')
 install_headers(lib_disk_io_headers, subdir : 'efimon/ebpf-modules/disk-io')
 install_headers(lib_power_headers, subdir : 'efimon/power')
 install_headers(lib_proc_headers, subdir : 'efimon/proc')
