--- subprojects/Sampling_by_PID/include/meson.build
@@ -1,1 +1,0 @@
-# vmlinux.h is generated in main meson.build
