--- subprojects/Sampling_by_PID/src/.gitignore
@@ -1,8 +1,0 @@
-# Auto-generated eBPF files (generated by meson during build)
-vmlinux.h
-prog.bpf.o
-prog.skel.h
-
-# Compiled binaries
-sampler
-benchmark
